moved serializer into same package a serialized class
